Added 64-bit Windows support. Updated existing windows sample projects and new project template to have x64 targets in addition to x86/Win32. Rebuilt all windows dependency libs and removed zlib.dll as zlib is now built as a static library. Note that any Windows projects generated with the old "new project script" will no longer link correctly with these changes since library paths have changed. Old projects will need to append /x86 to their windows lib paths.
Fixes #215

+++ gameplay/src/Game.inl
@@ -123,7 +123,7 @@ inline void Game::displayKeyboard(bool display)
 
 inline unsigned int Game::getGamepadCount() const
 {
-    return _gamepads->size();
+    return (unsigned int)_gamepads->size();
 }
 
 inline unsigned int Game::getGamepadsConnected()
